refactor(*): kebab-case all the things
