+++ config/eslint-config.json
@@ -1,5 +1,5 @@
 {
-  "plugins": ["security", "compat", "mocha", "import"],
+  "plugins": ["security", "compat", "mocha", "import", "unicorn"],
   "parserOptions": {
     "ecmaVersion": 6,
     "sourceType": "module",
@@ -302,8 +302,56 @@
     "import/max-dependencies": "error",
     "import/no-unassigned-import": "error",
     "import/no-named-default": "error",
-    "import/no-anonymous-default-export": "error"
-    // "import/group-exports": "error"
+    "import/no-anonymous-default-export": "error",
+    "import/group-exports": "error",
+
+    "unicorn/better-regex": "off",
+    "unicorn/catch-error-name": ["error", { "name": "pError" }],
+    "unicorn/consistent-function-scoping": "off",
+    "unicorn/custom-error-definition": "off",
+    "unicorn/error-message": "error",
+    "unicorn/escape-case": "error",
+    "unicorn/expiring-todo-comments": "error",
+    "unicorn/explicit-length-check": "error",
+    "unicorn/filename-case": ["error", { "case": "kebabCase" }],
+    "unicorn/import-index": "off",
+    "unicorn/new-for-builtins": "off", // conflicts with security/detect-non-literal-regexp, apparently
+    "unicorn/no-abusive-eslint-disable": "error",
+    "unicorn/no-array-instanceof": "error",
+    "unicorn/no-console-spaces": "error",
+    "unicorn/no-fn-reference-in-iterator": "off",
+    "unicorn/no-for-loop": "error",
+    "unicorn/no-hex-escape": "error",
+    "unicorn/no-keyword-prefix": "off",
+    "no-nested-ternary": "off",
+    "unicorn/no-nested-ternary": "error",
+    "unicorn/no-new-buffer": "error",
+    "unicorn/no-process-exit": "error",
+    "unicorn/no-unreadable-array-destructuring": "error",
+    "unicorn/no-unsafe-regex": "off", // off b/c we already use "security/detect-unsafe-regex"
+    "unicorn/no-unused-properties": "off",
+    "unicorn/no-zero-fractions": "error",
+    "unicorn/number-literal-case": "error",
+    "unicorn/prefer-add-event-listener": "error",
+    "unicorn/prefer-dataset": "error",
+    "unicorn/prefer-event-key": "error",
+    "unicorn/prefer-flat-map": "error",
+    "unicorn/prefer-includes": "error",
+    "unicorn/prefer-modern-dom-apis": "error",
+    "unicorn/prefer-negative-index": "error",
+    "unicorn/prefer-node-append": "error",
+    "unicorn/prefer-node-remove": "error",
+    "unicorn/prefer-query-selector": "error",
+    "unicorn/prefer-reflect-apply": "error",
+    "unicorn/prefer-replace-all": "off",
+    "unicorn/prefer-spread": "error",
+    "unicorn/prefer-starts-ends-with": "error",
+    "unicorn/prefer-string-slice": "error",
+    "unicorn/prefer-text-content": "error",
+    "unicorn/prefer-trim-start-end": "error",
+    "unicorn/prefer-type-error": "off",
+    "unicorn/prevent-abbreviations": "off",
+    "unicorn/throw-new-error": "error"
   },
   "env": {
     "node": true,
