+++ package.json
@@ -121,6 +121,7 @@
     "eslint-plugin-import": "2.20.1",
     "eslint-plugin-mocha": "6.3.0",
     "eslint-plugin-security": "1.4.0",
+    "eslint-plugin-unicorn": "15.0.1",
     "husky": "3.1.0",
     "lint-staged": "10.0.8",
     "mocha": "7.1.0",
