+++ src/cli/file-name-to-stream.js
