+++ src/cli/make-description.js
