+++ src/parse/scxml/normalize-machine.js
