+++ src/render/dot/index.js
@@ -1,7 +1,7 @@
 const Handlebars = require("handlebars/dist/handlebars.runtime");
 const _cloneDeep = require("lodash.clonedeep");
 const options = require("../../options");
-const StateMachineModel = require("../../stateMachineModel");
+const StateMachineModel = require("../../state-machine-model");
 const attributebuilder = require("./attributebuilder");
 const stateTransformers = require("./state-transformers");
 const transitionTransformers = require("./transition-transformers");
