+++ src/render/scjson/make-valid-event-names.js
