+++ src/render/scjson/make-valid-xml-name.js
