+++ src/state-machine-model.js
