+++ test/ast-massage-03-flatten-transitions.json
