+++ test/cli/attributes-parser.spec.js
@@ -11,8 +11,8 @@ function assertSyntaxError(pProgram, pParser, pErrorType) {
       lStillRan = true;
     }
     expect(lStillRan).to.equal(false);
-  } catch (e) {
-    expect(e.name).to.equal(pErrorType);
+  } catch (pError) {
+    expect(pError.name).to.equal(pErrorType);
   }
 }
 /*
