+++ test/transform/desugar-forks.spec.js
