+++ test/transform/desugar-joins.spec.js
