+++ test/transform/desugar-junctions.spec.js
