+++ utl/cut-handlebar-cookie.js
